[FEATURE] Retain scrollbar position after reloading
We save the scroll position only before refreshing the iframe (save/discard changes) and not when browsing the website directly because, in this last case, the user expect the scroll of the page to be on top. For the same reason we remove "scrollTop" from sessionStorage just after getting it.