[FEATURE][!!!] Allow any editable wrapper tag
ViewHelper core:contentEditable now accepts the argument `tag` (e.g. `tag="ul"`).

When frontend editing is not in use and a `tag` argument is specified, a tag of the specified type will be rendered.

This makes it possible to enable editing e.g. only for the content of an unordered list by making the <ul> tag the editor wrapper.