[BUGFIX] Dynamically add a spacer to the top of the page
This spacer is needed if there's not enough space between the top of the page and the first editable element to allow the inline actions toolbar of this first editable CE to be visible. Without the spacer, in this use case, the inline actions toolbar is hidden under the module doc header.

The code must be placed in .on('load') of inline_editing.css so the inline actions toolbar has its style applied and its height and position can be retrieved correctly.

Resolves: #728